requirements.txt
rich==13.7.1
python-dotenv
pydantic<2
langchain
langchain-openai
langchain-community
langserve[all]